istioctl: remove experimental `add-to-mesh`, `remove-from-mesh` and `kube-uninject`
Rationale:
* These have been experimental since inception many many years ago with
  no path to promotion
* They promote bad practices; usage of automatic injection is strongly
  preferred
* The VM part uses legacy VM model not documented/recommended anymore
* It doesn't really work. Uninjecting is not feasible; we just attempt
  to guess what was injected and what wasn't which is not possible to do
correctly.

As a result, I recommend removing these entirely.
