+++ tests/integration/pilot/main_test.go
@@ -20,12 +20,9 @@ package pilot
 import (
 	"testing"
 
-	"istio.io/istio/pkg/config/protocol"
 	"istio.io/istio/pkg/test/framework"
-	"istio.io/istio/pkg/test/framework/components/echo"
 	"istio.io/istio/pkg/test/framework/components/echo/common/deployment"
 	"istio.io/istio/pkg/test/framework/components/istio"
-	"istio.io/istio/pkg/test/framework/components/namespace"
 	"istio.io/istio/pkg/test/framework/resource"
 )
 
@@ -53,20 +50,4 @@ func TestMain(m *testing.M) {
 			return nil
 		}).
 		Run()
-}
-
-func echoConfig(ns namespace.Instance, name string) echo.Config {
-	return echo.Config{
-		Service:   name,
-		Namespace: ns,
-		Ports: []echo.Port{
-			{
-				Name:     "http",
-				Protocol: protocol.HTTP,
-				// We use a port > 1024 to not require root
-				WorkloadPort: 8090,
-			},
-		},
-		Subsets: []echo.SubsetConfig{{}},
-	}
 }
